removed unnecessary saving of raw and median data, changed save_image to float32, fixed the dimensions of the plot

+++ utils/misc_utils.py
@@ -119,7 +119,7 @@ def imshowComplex(vol, blocking=False):
 
 def save_image(tensor, path='output.png'):
     if 'tif' in path:
-        imsave(path, tensor[0, ...].cpu().numpy().astype(np.float16))
+        imsave(path, tensor[0, ...].cpu().numpy().astype(np.float32))
         return
     if tensor.shape[1] == 1:
         imshow2D(tensor)
